api/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlalchemy>=2.0
openai
python-dotenv
docxtpl
jinja2
